fix(web): move the suggestion source guards out of the server-action module
`lib/actions/prompt-suggestions.ts` is `'use server'`, where every runtime
export has to be an async server action. The two synchronous narrowing
helpers added with the Analytics source therefore failed the production
build with "export gscSourceData was not found" — while tsc, eslint and
prettier all passed, because none of them run the Next build.

The types and their guards now live in a plain module and the server-action
file imports the types alone.
